% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryJaya.R
\name{summary.jaya}
\alias{summary.jaya}
\title{Function to summarize the Jaya function}
\usage{
\method{summary}{jaya}(object, ...)
}
\arguments{
\item{object}{as an output object from 'jaya' function}

\item{\dots}{Additional parameters given to the function}
}
\value{
returns the summary of output object from 'jaya' function
}
\description{
Function to summarize the Jaya function
}
\examples{
# Test Function to minimize
square <- function(x){return((x[1]^2)+(x[2]^2))}
a <- jaya(fun = square, lower = c(-100,-100), upper = c(100,100), maxiter = 10, n_var = 2)
summary(a)
}
