% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotJaya.R
\name{plot.jaya}
\alias{plot.jaya}
\title{#' Function to plot the 'best value' VS 'no. of iterations'}
\usage{
\method{plot}{jaya}(x, ...)
}
\arguments{
\item{x}{as an output object from 'jaya' function}

\item{...}{as Additional graphical parameters given to plot function}
}
\value{
Returns plot showing 'best value' VS 'no. of iterations'
}
\description{
#' Function to plot the 'best value' VS 'no. of iterations'
}
